+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Added Settings
+- New setting: Start minimized
+- New setting: Minimize to tray on close
 - Added System Tray Icon
 - Added Hidden mode CLI Argument (-h|hidden)